Move brave sync switches under components/brave_sync
