+++ common/brave_switches.h
@@ -8,10 +8,6 @@
 
 namespace switches {
 
-// This function can be called from any thread, and the implementation doesn't
-// assume it's running on the UI thread.
-bool IsBraveSyncAllowedByFlag();
-
 // All switches in alphabetical order. The switches should be documented
 // alongside the definition of their values in the .cc file.
 extern const char kDisableBraveExtension[];
@@ -25,8 +21,6 @@ extern const char kDisableChromeGoogleURLTrackingClient[];
 extern const char kDisableTorClientUpdaterExtension[];
 
 extern const char kDisableWebTorrentExtension[];
-
-extern const char kDisableBraveSync[];
 
 extern const char kRewards[];
