+++ components/brave_sync/BUILD.gn
@@ -66,6 +66,17 @@ source_set("prefs") {
   ]
 }
 
+source_set("switches") {
+  sources = [
+    "switches.cc",
+    "switches.h",
+  ]
+
+  deps = [
+    "//base",
+  ]
+}
+
 source_set("public") {
   sources = [
     "public/brave_profile_sync_service.h",
@@ -75,13 +86,6 @@ source_set("public") {
     ":core",
     "buildflags",
   ]
-  if (!is_ios) {
-    deps += [
-      # For IsBraveSyncAllowedByFlag in
-      # components/sync/driver/sync_driver_switches.cc
-      "//brave/common:switches",
-    ]
-  }
 }
 
 source_set("core") {
@@ -101,6 +105,7 @@ source_set("core") {
   deps = [
     ":jslib_messages",
     ":prefs",
+    ":switches",
     "//base",
     "//components/bookmarks/browser",
     "//crypto",
