+++ components/brave_sync/switches.cc
@@ -3,17 +3,19 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#define IsSyncAllowedByFlag IsSyncAllowedByFlag_ChromiumImpl
-#include "../../../../../components/sync/driver/sync_driver_switches.cc"  // NOLINT
-#undef IsSyncAllowedByFlag
-
 #include "brave/components/brave_sync/switches.h"
 
+#include "base/command_line.h"
+
+namespace brave_sync {
 namespace switches {
 
-bool IsSyncAllowedByFlag() {
-  return IsSyncAllowedByFlag_ChromiumImpl() &&
-         brave_sync::switches::IsBraveSyncAllowedByFlag();
+bool IsBraveSyncAllowedByFlag() {
+  return !base::CommandLine::ForCurrentProcess()->HasSwitch(kDisableBraveSync);
 }
 
+// Allows disabling Brave Sync.
+const char kDisableBraveSync[] = "disable-brave-sync";
+
 }  // namespace switches
+}  // namespace brave_sync
